feat: Enhance device management and UI
Adds company/group assignment to devices and allows for editing this property within the DeviceList component. This change also refines the UI by updating the favicon, introducing a company logo in the sidebar and auth page, and improving search functionality in the DeviceList to include company names.

Additionally, several UI elements and icons were updated for a more polished look and feel.